Refactor files to clarify sample Executor
