+++ CMakeLists.txt
@@ -1001,7 +1001,7 @@ if(BUILD_TESTS)
 
   add_e2e_test(
     NAME external_executor_test
-    PYTHON_SCRIPT ${CMAKE_SOURCE_DIR}/tests/external_executor.py
+    PYTHON_SCRIPT ${CMAKE_SOURCE_DIR}/tests/external_executor/external_executor.py
     CONSENSUS ${CONSENSUS_FILTER}
   )
 endif()
