+++ src/apps/external_executor/protobuf/CMakeLists.txt
@@ -36,8 +36,8 @@ foreach(proto_file ${PROTO_FILES})
   )
 
   add_custom_command(
-    OUTPUT ${CMAKE_SOURCE_DIR}/tests/${PROTO_NAME_WE}_pb2.py
-           ${CMAKE_SOURCE_DIR}/tests/${PROTO_NAME_WE}_pb2_grpc.py
+    OUTPUT ${CMAKE_SOURCE_DIR}/tests/external_executor/${PROTO_NAME_WE}_pb2.py
+           ${CMAKE_SOURCE_DIR}/tests/external_executor/${PROTO_NAME_WE}_pb2_grpc.py
     COMMAND ${CMAKE_CURRENT_SOURCE_DIR}/build.sh ${proto_file}
             ${CMAKE_SOURCE_DIR}/tests/
     COMMENT "Generate Python source file from protobuf file ${PROTO_NAME}"
@@ -46,8 +46,8 @@ foreach(proto_file ${PROTO_FILES})
   )
   add_custom_target(
     ${PROTO_NAME_WE}_proto_python ALL
-    DEPENDS ${CMAKE_SOURCE_DIR}/tests/${PROTO_NAME_WE}_pb2.py
-            ${CMAKE_SOURCE_DIR}/tests/${PROTO_NAME_WE}_pb2_grpc.py
+    DEPENDS ${CMAKE_SOURCE_DIR}/tests/external_executor/${PROTO_NAME_WE}_pb2.py
+            ${CMAKE_SOURCE_DIR}/tests/external_executor/${PROTO_NAME_WE}_pb2_grpc.py
     JOB_POOL one_job
   )
